Adding scripts automate hadoop setup

+++ setup/lancluster/all-node.sh
@@ -2,28 +2,25 @@
 # Proper header for a Bash script.
 # Ref http://diybigdata.net/2017/11/upgrading-odroid-cluster-to-ubuntu-16-04/
 # Login as 'root' user
+if [ $(whoami) != 'root' ]; then echo "Please run as 'root' user"; exit 1; fi
 
 # Fix perl complaining error of some missing locales. See https://gist.github.com/panchicore/1269109
-#export LANGUAGE=en_US.UTF-8
-#export LANG=en_US.UTF-8
-#export LC_ALL=en_US.UTF-8
-locale-gen en_US.UTF-8
-#dpkg-reconfigure locales
+sudo locale-gen en_US.UTF-8
 
 # Create user 'odroid' in sudo group if not yet exists
 id -u odroid &>/dev/null || {
-	read -p "[*] Adding user 'odroid' as sudo user..."
+	read -p "[*] Add user 'odroid' as sudo user [Enter] "
 	adduser odroid
 	adduser odroid sudo
 }
 
-read -p "[*] Upgrading apts..."
+read -p "[*] Upgrade apts [Enter] "
 apt-get update
 apt-get upgrade
 apt autoremove
 
-rm /var/lib/dpkg/lock    
-rm /var/cache/apt/archives/lock
+sudo rm /var/lib/dpkg/lock    
+sudo rm /var/cache/apt/archives/lock
 
 # Switch to user 'odroid'
 #read -p "Switching to user 'odroid'..."
@@ -34,23 +31,40 @@ rm /var/cache/apt/archives/lock
 # 2. Turn off Xorg
 # 3. Resize the Boot Drive
 #sudo -s
-read -p "[*] Now installing odroid-utility.sh..."
-wget -O /usr/local/bin/odroid-utility.sh https://raw.githubusercontent.com/mdrjr/odroid-utility/master/odroid-utility.sh
-chmod +x /usr/local/bin/odroid-utility.sh
-read -p "[*] Modifying odroid-utility.sh, do these 3 things: Change hostname; Turn off Xorg; Resize the Boot Drive. Ready?..."
-odroid-utility.sh
+while true; do
+    read -p "[*] Install odroid-utility.sh? [y/n]: " yn
+    case $yn in
+        [Yy]* ) 
+        	wget -O /usr/local/bin/odroid-utility.sh https://raw.githubusercontent.com/mdrjr/odroid-utility/master/odroid-utility.sh
+			chmod +x /usr/local/bin/odroid-utility.sh
+			read -p "[*] Modifying odroid-utility.sh. Do these 3 things: Change hostname; Turn off Xorg; Resize the Boot Drive. Ready?..."
+			odroid-utility.sh
+        	break;;
+        [Nn]* ) break;;
+        * ) ;;
+    esac
+done
 
 # Override the whole /etc/hosts
-read -p "[*] Writing to /etc/hosts..."
+echo '------------------'
+cat hosts
+echo '------------------'
+read -p "[*] Write content above to /etc/hosts [Enter] "
 cat hosts > /etc/hosts
-# Append this to /etc/sysctl.conf
-read -p "[*] Writing to /etc/sysctl.conf..."
+
+# Append to /etc/sysctl.conf
+echo '------------------'
+cat sysctl.conf
+echo '------------------'
+read -p "[*] Appending content above to /etc/sysctl.conf [Enter] "
 cat sysctl.conf >> /etc/sysctl.conf
 
 # Log eth* MAC addr
-read -p "[*] Please log eth* MAC addr for later registering with master node for static IP assignment..."
+echo '------------------'
 ifconfig -a
+echo '------------------'
+read -p "[*] Please log eth* MAC addr for registering static IP addr with master node later [Enter] "
 
 # Done
-read -p "[*] Done. Now you can do: 'shutdown -h now'"
+read -p "[*] DONE [Enter] "
